requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.4
djangorestframework==3.14.0
idna==3.4
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
soupsieve==2.4.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
